SpiderMonkey bindings are more correct
The code runs :). However, all of the text is black.

+++ tealeaf/platform/iosVersioning.h
@@ -29,7 +29,7 @@
 #define SYSTEM_VERSION_LESS_THAN_OR_EQUAL_TO(v)     ([[[UIDevice currentDevice] systemVersion] compare:v options:NSNumericSearch] != NSOrderedDescending)
 
 
-CEXPORT int get_platform_memory_limit();
+CEXPORT long get_platform_memory_limit();
 
 CEXPORT NSString *get_platform();
